django/requirements.txt
amqp==5.2.0
argcomplete==3.5.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
billiard==4.2.0
black==24.8.0
boto3==1.34.149
botocore==1.34.149
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
commitizen==3.29.0
constantly==23.10.4
cryptography==42.0.8
daphne==4.0.0
decli==0.6.2
defusedxml==0.8.0rc2
Django==4.2.22
django-auth==0.1.9
django-cors-headers==4.4.0
django-environ==0.11.2
django-oauth-toolkit==2.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==2.3.0
drf-yasg==1.21.7
et_xmlfile==2.0.0
firebase-admin==6.6.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
httpx==0.27.2
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.2.0
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
jwcrypto==1.5.6
kombu==5.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.0.8
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==2.1.2
oauthlib==3.2.2
openai==1.51.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prompt-toolkit==3.0.36
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyparsing==3.1.4
pytest==8.3.3
pytest-django==4.9.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
questionary==2.0.1
redis==5.0.1
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==2.0.0
aiohttp==3.10.5
psutil==6.0.0
rsa==4.9
s3transfer==0.10.2
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
SQLAlchemy==2.0.36
sqlparse==0.5.0
sympy==1.13.3
termcolor==2.4.0
tomli==2.0.1
tomlkit==0.13.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
watchdog==5.0.3
wcwidth==0.2.13
zipp==3.20.0
zope.interface==7.0.1
